feat: refactor how actions are created
